+++ python/ry/ryo3/_http.pyi
